PRO-3555: Add/Amend unit and component test and remove leftover rogue code
