+++ test/component/testThankYouPage.js
@@ -1,0 +1,36 @@
+'use strict';
+
+const TestWrapper = require('test/util/TestWrapper');
+const config = require('app/config');
+
+describe('thank-you', () => {
+    let testWrapper;
+
+    beforeEach(() => {
+        testWrapper = new TestWrapper('ThankYou');
+    });
+
+    afterEach(() => {
+        testWrapper.destroy();
+    });
+
+    describe('Verify Content, Errors and Redirection', () => {
+        it('test content loaded on the page', (done) => {
+            const sessionData = {
+                ccdCase: {
+                    id: '1234-5678-9012-3456'
+                }
+            };
+            testWrapper.agent.post('/prepare-session/form')
+                .send(sessionData)
+                .end(() => {
+                    const contentData = {
+                        helpLineNumber: config.helpline.number,
+                        findOutNext: config.links.findOutNext
+                    };
+
+                    testWrapper.testContent(done, [], contentData);
+                });
+        });
+    });
+});
